refactor: DateFormatter extension에 format(date:)을 타입메서드로 구현
